… src/test/components/__snapshots__/MarkerChart.test.js.snap
… src/test/components/__snapshots__/MarkerTable.test.js.snap
… src/test/store/__snapshots__/profile-view.test.js.snap