[This profile](https://perfht.ml/2G0dhpa) has a large amount of markers and is slow on rendering the marker chart. I switched it to use the `getRangeSelectionFilteredTracingMarkers` and things got fast again. The tests need fixing for this, so it's still WIP.

STR: Mouse over the markers in the marker chart.

Before:
```
[timing]    "markerChartCanvas render" took 3068ms to execute.
[timing]    "markerChartCanvas render" took 958ms to execute.
[timing]    "markerChartCanvas render" took 588ms to execute.
[timing]    "markerChartCanvas render" took 222ms to execute.
[timing]    "markerChartCanvas render" took 220ms to execute.
[timing]    "markerChartCanvas render" took 104ms to execute.
[timing]    "markerChartCanvas render" took 188ms to execute.
[timing]    "markerChartCanvas render" took 346ms to execute.
```

After:
```
[timing]    "markerChartCanvas render" took 12ms to execute.
[timing]    "markerChartCanvas render" took 10ms to execute.
[timing]    "markerChartCanvas render" took 8ms to execute.
[timing]    "markerChartCanvas render" took 4ms to execute.
[timing]    "markerChartCanvas render" took 6ms to execute.
[timing]    "markerChartCanvas render" took 10ms to execute.
[timing]    "markerChartCanvas render" took 6ms to execute.
[timing]    "markerChartCanvas render" took 8ms to execute.
[timing]    "markerChartCanvas render" took 6ms to execute.
[timing]    "markerChartCanvas render" took 8ms to execute.
[timing]    "markerChartCanvas render" took 6ms to execute.
[timing]    "markerChartCanvas render" took 8ms to execute.```
```