@gregtatum 2018-05-30
@julienw This is ready for review. I added a test for it. I also don't think the perf of running the selectors is critical as it's not on a hot path, and doesn't feel slow when I use the tool. I didn't include a profile of that as you commented.