@julienw 2018-07-13
Ah nice ! I don't remember why I missed it. I even feel like I looked at it already... anyway this is in my TODO now.