@julienw 2018-07-16
**I took some measurements with #1134:**

* Loading the profile: https://perfht.ml/2LjtyVM (Note: I did it several times and didn't always get the same measurements, so take with a grain of salt; this one is after minimizing the memory)
* Hovering some markers to display their tooltips: https://perfht.ml/2LnusRf (zoomed in on one of the biggest hang -- slightly more than 100ms)
* Selecting a range (but not "committing" it) from the "2.7sec" range: https://perfht.ml/2uCBomo
* Moving back from 6ms to 2.7sec: https://perfht.ml/2LnLf6L
  I expect this is where this PR can change things for the worse.
* Moving back from 6ms to full range: https://perfht.ml/2uEVvA6
  Same consideration.
* Moving from full range to 6ms: https://perfht.ml/2uFnccb

The main benefit without this patch is that the MarkerTimings need to be computed only once at load time. The downside is that we need to compute them for all markers even if we display a part of it.

**And now with both #1134 and this patch:**

* Loading the profile: https://perfht.ml/2LlyxoO
   `getMarkerTiming` is now nearly absent, which was expected.
* Hovering some markers to display their tooltips: https://perfht.ml/2LhF6Mo
  We definitely see that `drawMarkers` is a lot less visible
* Selecting a range: https://perfht.ml/2LpSVFl
  Same here.
* Moving back from 6ms to 2.7sec: https://perfht.ml/2Lk4PRm
   I'm very surprised that `drawOneMarker` is much more visible. I retried this scenario several times, with and without the patch, but this is pretty consistent. This is as if we're drawing twice as much, but I don't see how this is possible... (unless I screwed up my rebase in some way ?)
* Moving back from 6ms to full range: https://perfht.ml/2LjbhYD
  We see the same thing happening. Also obviously `getMarkerTiming` needs to be recomputed, which takes time in this case. But if we can fix the issue with `drawOneMarker` I think this is acceptable.
* Moving from full range to 6ms: https://perfht.ml/2uE8vGn
  This one looks worse too, but I think this isn't significant.

When using it for real there doesn't seem to be very different performance characteristics with this PR. Although I'd like to understand why `drawOneMarker` is called more (?) for some scenarios, maybe there _is_ something (else) fishy happening here.

**UX difference**

In terms of UX I prefer the behavior with this patch. Compare without the patch:
![Lots of empty lines](https://user-images.githubusercontent.com/454175/42772434-b5c1f0c2-892a-11e8-9da6-09ca97fa8999.png)

with the patch:
![no empty lines !](https://user-images.githubusercontent.com/454175/42772440-bd517a10-892a-11e8-9eec-de5e03e41e16.png)


So I find the second behavior much cleaner.